+++ saleor/extensions/__init__.py
@@ -1,6 +1,8 @@
 import importlib
 from typing import List
 
+from django.utils.translation import pgettext_lazy
+
 from .checks import check_extensions  # NOQA: F401
 
 
@@ -17,3 +19,15 @@ def discover_plugins_modules(plugins: List[str]):
         module = importlib.import_module(module_path)
         plugins_modules.append(module.__package__)
     return plugins_modules
+
+
+class ConfigurationTypeField:
+    STRING = "String"
+    BOOLEAN = "Boolean"
+    CHOICES = [
+        (STRING, pgettext_lazy("Type of the configuration field", "Field is a String")),
+        (
+            BOOLEAN,
+            pgettext_lazy("Type of the configuration field", "Field is a Boolean"),
+        ),
+    ]